+++ backend/src/main/java/edu/netcracker/jobdealer/repository/SkillsRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface SkillsRepository extends JpaRepository<Skills, Integer> {
+public interface SkillsRepository extends JpaRepository<Skills, Long> {
     Optional<Skills> findByName(String skill);
 
 }
